[sc-8798] Check sync server status
- Check sync server status every 5s when it is down, and every 5min when it is up
- Push a notification when the server is down
- Move notifications folder to `core` lib to prevent circle dependencies
