+++ libs/common/src/assets/i18n/fr.json
@@ -629,6 +629,7 @@
 	"notification.type.deletion-failed": "Échec de la suppression de ressource",
 	"notification.type.modification-completed": "Édition de la ressource terminée",
 	"notification.type.modification-failed": "Échec de l'édition de la ressource",
+	"notification.type.sync-server-down": "Le serveur de synchronisation est en panne",
 	"pastanaga.calendar.invalid-format": "Format de date non valide (format attendu : \"MMM dd, YYYY\")",
 	"pastanaga.calendar.months": "Mois",
 	"pastanaga.calendar.years": "Années",
