+++ libs/core/src/lib/notifications/index.ts
@@ -2,3 +2,4 @@ export * from './notification-button/notification-button.component';
 export * from './notifications-panel/notifications-panel.component';
 export * from './notification/notification.component';
 export * from './notification.model';
+export * from './notification.service';
