+++ libs/core/src/lib/notifications/notification.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { BehaviorSubject, combineLatest, filter, map, Observable, switchMap, tap } from 'rxjs';
 import { NotificationData, NotificationUI } from './notification.model';
 import { SDKService } from '@flaps/core';
-import { NavigationService } from '../services';
+import { NavigationService } from '../../../../common/src/lib/services';
 import { differenceInSeconds } from 'date-fns';
 import { WritableKnowledgeBox } from '@nuclia/core';
 import { takeUntil } from 'rxjs/operators';
@@ -76,6 +76,10 @@ export class NotificationService {
         takeUntil(this.sdk.nuclia.auth.isAuthenticated().pipe(filter((isAuthenticated) => !isAuthenticated))),
       )
       .subscribe();
+  }
+
+  pushNotifications(notifications: NotificationUI[]) {
+    this._notifications.next(notifications.concat(this._notifications.value));
   }
 
   markAllAsRead() {
