+++ libs/core/src/lib/notifications/notification/notification.component.scss
@@ -1,4 +1,4 @@
-@import 'apps/dashboard/src/variables';
+@import '../../../../../../apps/dashboard/src/variables';
 
 .notification-container {
   padding: rhythm(1) rhythm(2);
