+++ libs/core/src/lib/notifications/notifications-panel/notifications-panel.component.scss
@@ -1,4 +1,4 @@
-@import 'apps/dashboard/src/variables';
+@import '../../../../../../apps/dashboard/src/variables';
 
 $notification-panel-width: rhythm(56);
